Add PostgreSQL test, update CircleCI config, improve verbose option
Not gonna lie, this was a real ballache: looks like PG 13 has a very
subtle bug in it. When you create a primary key, it's supposed to
be recorded in the constraint_column_usage view in the information_schema.

And it does show up there, BUT ONLY IF YOU CREATED THE DATABASE! Even if
you're an admin user, it doesn't show up there.

Eventually I worked out that you can use key_column_usage instead.

How lucky am I? First time I've ever used PostgreSQL and I finnd a bug!